[brcm47xx]: activate failsafe on all ports of the switch and not just port 0.
Sometimes the port 0 was the WAN port or there could be some device 
where the port 0 is not wired to an actual Ethernet port. Now telnet in 
failsafe will listen on all ports of the switch so user do not have to 
search for the correct port any more. This change will make the switch 
leak traffic between the WAN and the LAN port, but you should not use 
failsafe while connected to a untrusted network.


git-svn-id: svn://svn.openwrt.org/openwrt/trunk@33544 3c298f89-4303-0410-b956-a3cf2f4a3e73