toolchain: stop generating broken symlink in TOOLCHAIN_DIR/lib
It turns out that the symlink may exists already by the time the toolchain
Makefile gets an opportunity to run. While we asked to replace the target, ln
doesn't do so until after dereferencing the existing symlink. This results in
an unintended symlink that refers to itself. Instead, create the link without
dereferencing any symlinks by using the -n option.

Signed-off-by: John Szakmeister <john@szakmeister.net>

git-svn-id: svn://svn.openwrt.org/openwrt/trunk@36140 3c298f89-4303-0410-b956-a3cf2f4a3e73